requirements.txt
# Core dependencies
requests>=2.31.0
pandas>=2.0.0
# pandas_ta requires numba which doesn't support Python 3.14 yet
# Using ta-lib alternative or manual indicator calculations
# pandas_ta>=0.3.14b
ta>=0.11.0
fastparquet>=2023.10.0
pyarrow>=14.0.0
tqdm>=4.66.0

# Environment management
python-dotenv>=1.0.0

# Development
pytest>=7.4.0
